refactor(ducklake): fix circular import
